feat(live-edit): P4 — need colour, effective clamp, live rate; and a 48-bit fix
needs_eval::need_state returns (value, min, max, rate) from ONE evaluation: the
bounds are the EFFECTIVE clamp, and the rate is segment_rate sampled at the
current elapsed — the same instantaneous rate the integrator averages, not a
second implementation. Sign flipped at the source so negative always means
losing satisfaction, and the panel colours blindly.

The rate needed a unit: per-tic is honest and unreadable (0.0023/tic shows as
-0.0). The panel formats per HOUR because that is what the corpus authors in —
thirst reads -100/h against its deplete = 21600, hunger -50.0/h against 43200.

FOUND AND FIXED (I9): MoverLayer.pawnNeeds flattened 48-bit need rows into a
Uint32Array, cutting off the data half — the VALUE. Every need read 0 client
side, which meant band-derived conditions could NEVER fire: the same pawn that
showed Scared/Uncomfortable now shows Dehydrated/Quenched. It was also the root
of all 7 pre-existing type errors; Float64Array takes the client typecheck to
ZERO errors for the first time this session.

Kept in this stream rather than deferred because it blocked P4 — a needs tab
that reads every need as 0 is not a needs tab.

+++ client/webgl/src/game/world/MoverLayer.ts
@@ -615,11 +615,20 @@ export class MoverLayer {
   }
 
   /** The entity's `needs` rows, flattened stride-2 `[packed, setTic, …]` (stat-model F2) —
-   *  the second eval input; fed verbatim to `pawnConditions`/`pawnEmotion`. */
-  pawnNeeds(entity: number): Uint32Array {
+   *  the second eval input; fed verbatim to `pawnConditions`/`pawnEmotion`/`needState`.
+   *
+   *  **Float64Array, not Uint32Array.** A need row is the ONE u64 shape
+   *  `dead:16 | data:16 | reference:32` (trait-rows-u32 F1) — 48 significant bits, which a JS
+   *  number holds exactly under THE 48-BIT LAW but a `Uint32Array` silently truncates. It used
+   *  to be `Uint32Array`, which cut `0xa4fb80010020` down to `0x80010020` and threw away the
+   *  `data` half — i.e. the need's VALUE. Every consumer therefore read every need as 0, so
+   *  band-derived conditions (thirsty / hungry / starving) could never fire on the client and
+   *  the live-edit needs bars were all empty. Found 2026-08-09 by probing the raw rows when
+   *  those bars read zero on a healthy pawn (live-edit I9). */
+  pawnNeeds(entity: number): Float64Array {
     const rows = this.needRows.get(entity);
-    if (!rows) return new Uint32Array(0);
-    const out = new Uint32Array(rows.size * 2);
+    if (!rows) return new Float64Array(0);
+    const out = new Float64Array(rows.size * 2);
     let i = 0;
     for (const [, [packed, setTic]] of rows) {
       out[i++] = packed;
